+++ src/main/kotlin/io/snyk/plugin/snykcode/core/SnykCodeParams.kt
@@ -1,6 +1,7 @@
 package io.snyk.plugin.snykcode.core
 
 import ai.deepcode.javaclient.core.DeepCodeParamsBase
+import com.intellij.openapi.diagnostic.Logger
 import io.snyk.plugin.getWaitForResultsTimeout
 import io.snyk.plugin.pluginSettings
 import io.snyk.plugin.snykcode.codeRestApi
@@ -19,10 +20,30 @@ class SnykCodeParams private constructor() : DeepCodeParamsBase(
     { getWaitForResultsTimeout() },
     codeRestApi
 ) {
+    private fun requestLogging() = Logger.getInstance(SCLogger.presentableName + "RequestLogging").isDebugEnabled
 
     init {
-        val apiUrl = toSnykCodeApiUrl(pluginSettings().customEndpointUrl)
-        setApiUrl(apiUrl, pluginSettings().ignoreUnknownCA)
+        setApiUrl(
+            toSnykCodeApiUrl(pluginSettings().customEndpointUrl),
+            pluginSettings().ignoreUnknownCA,
+            requestLogging()
+        )
+    }
+
+    override fun setApiUrl(apiUrl: String) {
+        setApiUrl(
+            toSnykCodeApiUrl(apiUrl),
+            pluginSettings().ignoreUnknownCA,
+            requestLogging()
+        )
+    }
+
+    override fun setApiUrl(apiUrl: String, disableSslVerification: Boolean) {
+        setApiUrl(
+            toSnykCodeApiUrl(apiUrl),
+            disableSslVerification,
+            requestLogging()
+        )
     }
 
     override fun consentGiven(project: Any): Boolean {