+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## [2.4.28]
 
+### Changed
+- Enable HTTP request logging when Snyk Code logger in debug mode
+
 ### Fixed
 - generic .dcignore file creation failure and exceptions
+
 
 ## [2.4.27]
 
@@ -20,7 +24,6 @@
 
 - Updated Snyk Code API to 2.3.1 (limit file size to 1 MB)
 - Provide link to [Privacy Policy](https://snyk.io/policies/privacy/) and [Terms of Service](https://snyk.io/policies/terms-of-service/) on Welcome screen.
-
 ### Fixed
 
 - Avoiding IncorrectOperationException by remove direct project service's requests.